# [Opt] Replace "simplify" with "full_simplify" in "Simplified I"